refactor: crawler 하위에 stock 패키지 제거
