+++ src/main/java/com/bookspot/batch/step/processor/crawler/StockFilePathParser.java
@@ -1,4 +1,4 @@
-package com.bookspot.batch.step.processor.crawler.stock;
+package com.bookspot.batch.step.processor.crawler;
 
 import com.bookspot.batch.global.crawler.common.exception.ElementNotFoundException;
 import com.bookspot.batch.global.crawler.naru.CsvFilePath;
